Improve the structure of NewKeystoneTx to take in a var args header struct + cleanup
