+++ client/src/core/requests.ts
@@ -1,12 +1,8 @@
 // api requests types
 
 import { api } from 'core/config';
-i
-import {ethers, Signature} from 'ethers';
-import {ECDSASignature, ecsign} from 'ethereumjs-util';
-import {Buffer} from "buffer";
-import {isNil} from "lodash";
-import {NewKeystoneTx} from "./middleware";
+import {NewKeystoneTx} from "./middleware/middleware";
+import {WithECDSAAuth} from "./middleware/ecdsaPublicKeyAuth";
 
 export const ECDSAPublicKeyAuthHeader = "ecdsaPublicKeyAuth"
 
@@ -27,13 +23,13 @@ export interface MoveRequest {
 // api requests
 
 export const CreatePlayer = async (request: CreatePlayerRequest) => {
-  return api.post('/player', NewKeystoneTx(request, null));
+  return api.post('/player', NewKeystoneTx(request));
 };
 
 export const Fire = async (request: CreateProjectileRequest) => {
-  return api.post('/fire', NewKeystoneTx(request));
+  return api.post('/fire', NewKeystoneTx(request, WithECDSAAuth(request)));
 };
 
 export const Move = async (request: MoveRequest) => {
-  return api.post('/move', NewKeystoneTx(request));
+  return api.post('/move', NewKeystoneTx(request, WithECDSAAuth(request)));
 };
